# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vendorkart B2B Multi-Vendor Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get cart
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemRequest"
      responses:
        "200":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear cart
      responses:
        "200":
          description: Cart cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /cart/{itemId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemRequest"
      responses:
        "200":
          description: Item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for current user
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /payments/initiate:
    post:
      operationId: initiatePayment
      tags: [payments]
      summary: Initiate dummy payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiatePaymentRequest"
      responses:
        "200":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSession"
  /payments/confirm:
    post:
      operationId: confirmPayment
      tags: [payments]
      summary: Confirm payment (simulate success)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmPaymentRequest"
      responses:
        "200":
          description: Payment confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResult"
...
  /admin/orders:
    get:
      operationId: listAllOrders
      tags: [admin]
      summary: List all orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: All orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        icon:
          type: string
        image:
          type: string
        productCount:
          type: integer
        parentId:
          type: integer
      required:
        - id
        - name
        - slug
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        image:
          type: string
        parentId:
          type: integer
      required:
        - name
...
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        total:
          type: number
        itemCount:
          type: integer
      required:
        - items
        - subtotal
        - total
    CartItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: string
        vendorId:
          type: integer
        vendorName:
          type: string
        price:
          type: number
        quantity:
          type: integer
        moq:
          type: integer
        subtotal:
          type: number
      required:
        - id
        - productId
        - price
        - quantity
    CartItemRequest:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required:
        - productId
        - quantity
    UpdateCartItemRequest:
      type: object
      properties:
        quantity:
          type: integer
      required:
        - quantity
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled, refunded]
        paymentStatus:
          type: string
          enum: [pending, paid, failed, refunded]
        paymentMethod:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        shippingAddress:
          $ref: "#/components/schemas/Address"
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - orderNumber
        - status
        - paymentStatus
        - total
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: string
        vendorId:
          type: integer
        vendorName:
          type: string
        quantity:
          type: integer
        price:
          type: number
        subtotal:
          type: number
      required:
        - id
        - productId
        - quantity
        - price
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
      required:
        - orders
        - total
    CreateOrderRequest:
      type: object
      properties:
        shippingAddressId:
          type: integer
        couponCode:
          type: string
        notes:
          type: string
      required:
        - shippingAddressId
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    InitiatePaymentRequest:
      type: object
      properties:
        orderId:
          type: integer
        amount:
          type: number
        currency:
          type: string
      required:
        - orderId
        - amount
    PaymentSession:
      type: object
      properties:
        sessionId:
          type: string
        orderId:
          type: integer
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
        dummyOrderId:
          type: string
        expiresAt:
          type: string
          format: date-time
      required:
        - sessionId
        - orderId
        - amount
    ConfirmPaymentRequest:
      type: object
      properties:
        sessionId:
          type: string
        orderId:
          type: integer
      required:
        - sessionId
        - orderId
    PaymentResult:
      type: object
      properties:
        success:
          type: boolean
        transactionId:
          type: string
        orderId:
          type: integer
        amount:
          type: number
        message:
          type: string
      required:
        - success
        - transactionId
...
    Address:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        phone:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        country:
          type: string
        isDefault:
          type: boolean
      required:
        - id
        - addressLine1
        - city
        - state
        - pincode
...